back-end/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2.4
django-allauth==65.10.0
django-browser-reload==1.18.0
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-htmx==1.23.2
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
faiss-cpu==1.11.0.post1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.3
groq==0.30.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.70
langchain-groq==0.3.6
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langsmith==0.4.8
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
openai==0.28.0
orjson==3.11.0
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
propcache==0.3.2
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.9.0
PyMuPDF==1.26.3
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
triton==3.3.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
zstandard==0.23.0
